docs: remove outdated commit hashes from readme
these hashes are outdated and not recommended. i could update to the latest prior commit but that would also quickly be outdated. by forcing people to choose a hash, they are more likely to get the latest verison.

It's not practical to try to maintain these hashes as the library keeps evolving. maybe long term this could be dynamically loaded somehow, but it's not worth the time to figure that out right now